requirements.txt
async-timeout==4.0.3
asyncio==4.0.0
bidict==0.23.1
biosppy==2.2.3
bleak==0.22.3
blinker==1.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
colorlog==6.10.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.1.5
dbus-fast==2.24.4
dbus-next==0.2.3
dbus-python==1.4.0
easydev==0.13.3
Flask==3.0.3
fonttools==4.57.0
future==1.0.0
# h5py @ file:///croot/h5py_1715094734337/work
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
line_profiler==4.2.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mock==5.2.0
neurokit2==0.2.4
nolds==0.5.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
PeakUtils==1.3.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
ptyprocess==0.7.0
pycairo==1.21.0
pygame==2.6.1
PyGObject==3.48.1
pyhrv==0.4.1
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
shortuuid==1.0.13
six==1.17.0
spectrum==0.9.0
spicy==0.16.0
threadpoolctl==3.5.0
transitions==0.9.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
